3rdparty: update {bx,bimg,bgfx}

+++ 3rdparty/bx/src/semaphore.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2017 Branimir Karadzic. All rights reserved.
+ * Copyright 2010-2018 Branimir Karadzic. All rights reserved.
  * License: https://github.com/bkaradzic/bx#license-bsd-2-clause
  */
 
@@ -78,7 +78,7 @@ namespace bx
 
 		dispatch_time_t dt = 0 > _msecs
 			? DISPATCH_TIME_FOREVER
-			: dispatch_time(DISPATCH_TIME_NOW, _msecs*1000000)
+			: dispatch_time(DISPATCH_TIME_NOW, int64_t(_msecs)*1000000)
 			;
 		return !dispatch_semaphore_wait(si->m_handle, dt);
 	}
@@ -98,13 +98,13 @@ namespace bx
 
 	void toTimespecMs(timespec& _ts, int32_t _msecs)
 	{
-		toTimespecNs(_ts, _msecs*1000000);
+		toTimespecNs(_ts, uint64_t(_msecs)*1000000);
 	}
 
 	void add(timespec& _ts, int32_t _msecs)
 	{
 		uint64_t ns = toNs(_ts);
-		toTimespecNs(_ts, ns + _msecs*1000000);
+		toTimespecNs(_ts, ns + uint64_t(_msecs)*1000000);
 	}
 
 	Semaphore::Semaphore()
